refactor: registry/types.ts を shared/command_types.ts に移設
コマンド型定義（GuildChatInputCommand 等）は features と registry の
両方が依存する「契約」であるため shared/ に移設し、
features → registry の逆方向依存 35 件を解消する。
