+++ src/features/pick/pick_command.ts
@@ -6,7 +6,7 @@ import {
 
 import { handlePick } from '@/features/pick/pick';
 
-import type { GlobalChatInputCommand } from '@/registry/types';
+import type { GlobalChatInputCommand } from '@/shared/command_types';
 
 const pick = new SlashCommandBuilder()
     .setName('pick')
