+++ src/features/recruit/interactions/edit_recruit/recruit_editor_command.ts
@@ -2,7 +2,7 @@ import { ApplicationCommandType, ContextMenuCommandBuilder } from 'discord.js';
 
 import { createRecruitEditor } from '@/features/recruit/interactions/edit_recruit/recruit_editor';
 
-import type { MessageContextMenuCommand } from '@/registry/types';
+import type { MessageContextMenuCommand } from '@/shared/command_types';
 
 const recruitEditor = new ContextMenuCommandBuilder()
     .setName('募集を編集')
