+++ src/gateway/command_handler.ts
@@ -7,7 +7,7 @@ import { sendErrorLogs } from '@/infra/logging/send_error_logs';
 import { chatInputCommands } from '@/registry/command_registry';
 import { exists } from '@/shared/assert';
 
-import type { GlobalChatInputCommand, GuildChatInputCommand } from '@/registry/types';
+import type { GlobalChatInputCommand, GuildChatInputCommand } from '@/shared/command_types';
 
 const logger = log4js_obj.getLogger('interaction');
