+++ test/gateway/command_handler.test.ts
