+++ test/infra/external/splatoon3-ink/splatoon3_ink.test.ts
