+++ test/registry/register.test.ts
